requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0

# LLM Providers
google-generativeai>=0.3.0
anthropic>=0.18.0
openai>=1.12.0

# CLI/Terminal
rich>=13.0.0
click>=8.1.0
questionary>=2.0.0

# Orchestration Frameworks
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
pyautogen>=0.2.0

# Utilities
pyyaml>=6.0
jsonschema>=4.17.0
tenacity>=8.2.0  # Retry logic
tqdm>=4.65.0  # Progress bars

# PDF Processing
pymupdf>=1.23.0  # PDF annotation and manipulation

# Optional but recommended
pypdf>=3.17.0  # PDF text extraction fallback
pillow>=10.0.0  # Image processing
PyMuPDF>=1.23.0  # PDF annotation (fitz)
pymupdf>=1.23.0  # PDF annotation and manipulation

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
